+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/PlatformImpl.java
@@ -95,7 +95,7 @@ public final class PlatformImpl extends AbstractPlatform {
 
     @Override
     public boolean isKeyDown(final KeyMapping keyMapping) {
-        return InputConstants.isKeyDown(
+        return !keyMapping.isUnbound() && InputConstants.isKeyDown(
             Minecraft.getInstance().getWindow().getWindow(),
             ((KeyMappingAccessor) keyMapping).getKey().getValue()
         );