+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/packet/PacketIds.java
@@ -14,6 +14,7 @@ public final class PacketIds {
     public static final ResourceLocation GRID_INSERT = createIdentifier("grid_insert");
     public static final ResourceLocation GRID_EXTRACT = createIdentifier("grid_extract");
     public static final ResourceLocation GRID_SCROLL = createIdentifier("grid_scroll");
+    public static final ResourceLocation CRAFTING_GRID_CLEAR = createIdentifier("crafting_grid_clear");
     public static final ResourceLocation RESOURCE_FILTER_SLOT_UPDATE = createIdentifier("resource_filter_slot_update");
     public static final ResourceLocation RESOURCE_FILTER_SLOT_CHANGE = createIdentifier("resource_filter_slot_change");
     public static final ResourceLocation RESOURCE_FILTER_SLOT_AMOUNT_CHANGE =