+++ refinedstorage2-platform-forge/src/main/java/com/refinedmods/refinedstorage2/platform/forge/PlatformImpl.java
@@ -93,7 +93,7 @@ public final class PlatformImpl extends AbstractPlatform {
 
     @Override
     public boolean isKeyDown(final KeyMapping keyMapping) {
-        return InputConstants.isKeyDown(
+        return !keyMapping.isUnbound() && InputConstants.isKeyDown(
             Minecraft.getInstance().getWindow().getWindow(),
             keyMapping.getKey().getValue()
         );